Move EmbedDisclaimer component
