+++ apps/web/components/real.ts
@@ -95,4 +95,3 @@ export * from './CardWithFeaturedItems/CardWithFeaturedItems'
 export * from './GenericList/GenericList'
 export * from './GenericList/LatestGenericListItems/LatestGenericListItems'
 export * from './connected/ParentalLeaveCalculator/ParentalLeaveCalculator'
-export * from './EmbedDisclaimer/EmbedDisclaimer'
